Support building the code with Visual Studio 2010
Solution files and other project files related to VS2010 added as well.

+++ treefileloader.h
@@ -10,7 +10,7 @@ class TreeFileLoader
 public:
     TreeFileLoader( GenealogicalTree* tree );
 
-    int64_t LoadFile( const std::string& path );
+    long long LoadFile( const std::string& path );
 
     std::string GetLastError();
